Refactored `rustcoalescence/src/args/mod.rs`
